requirements.txt
getmac==0.9.4
mysql-connector-python==8.4.0
netaddr==1.2.1
zeep==4.2.1
pigpio==1.78
paho_mqtt==2.1.0
psutil==6.0.0
syslog-rfc5424-formatter==1.2.3